+++ draft-packages/modal/KaizenDraft/Modal/Presets/InputEditModal.tsx
@@ -1,8 +1,8 @@
-import classnames from "classnames"
 import * as React from "react"
 
 import { Text } from "@kaizen/component-library"
 
+import { ButtonProps } from "@kaizen/draft-button"
 import {
   GenericModal,
   ModalAccessibleLabel,
@@ -13,7 +13,7 @@ import {
 
 import styles from "./InputEditModal.scss"
 
-interface Props {
+export interface InputEditModalProps {
   readonly isOpen: boolean
   readonly type: "positive" | "negative"
   readonly title: string
@@ -24,54 +24,68 @@ interface Props {
   readonly dismissLabel?: string
   readonly automationId?: string
   readonly children: React.ReactNode
-  readonly submitDisabled?: boolean
+  readonly submitWorking?: { label: string; labelHidden?: boolean }
 }
 
-type InputEditModal = React.FunctionComponent<Props>
+type InputEditModal = React.FunctionComponent<InputEditModalProps>
 
 const InputEditModal = ({
   isOpen,
   type,
   title,
   onSubmit,
-  onDismiss,
   localeDirection = "ltr",
   submitLabel = "Submit",
   dismissLabel = "Cancel",
+  submitWorking,
   automationId,
   children,
-  submitDisabled = false,
-}: Props) => (
-  <GenericModal
-    isOpen={isOpen}
-    onEscapeKeyup={onDismiss}
-    automationId={automationId}
-  >
-    <div className={styles.modal} dir={localeDirection}>
-      <ModalHeader unpadded onDismiss={onDismiss}>
-        <div className={styles.header}>
-          <ModalAccessibleLabel>
-            <Text tag="h1" style="zen-heading-3" inline>
-              {title}
-            </Text>
-          </ModalAccessibleLabel>
-        </div>
-      </ModalHeader>
-      <ModalBody unpadded>
-        <div className={styles.body} dir={localeDirection}>
-          {children}
-        </div>
-      </ModalBody>
-      <ModalFooter
-        actions={[
-          { label: submitLabel, action: onSubmit, disabled: submitDisabled },
-          { label: dismissLabel, action: onDismiss },
-        ]}
-        appearance={type === "negative" ? "destructive" : "primary"}
-        automationId={automationId}
-      />
-    </div>
-  </GenericModal>
-)
+  ...props
+}: InputEditModalProps) => {
+  const onDismiss = submitWorking ? undefined : props.onDismiss
+  const submitAction = { label: submitLabel, onClick: onSubmit }
+  const workingProps = submitWorking
+    ? {
+        working: true,
+        workingLabel: submitWorking.label,
+        workingLabelHidden: submitWorking.labelHidden,
+      }
+    : {}
+
+  const footerActions: ButtonProps[] = [
+    { ...submitAction, ...workingProps },
+    { label: dismissLabel, onClick: onDismiss, disabled: !!submitWorking },
+  ]
+
+  return (
+    <GenericModal
+      isOpen={isOpen}
+      onEscapeKeyup={onDismiss}
+      automationId={automationId}
+    >
+      <div className={styles.modal} dir={localeDirection}>
+        <ModalHeader unpadded onDismiss={onDismiss}>
+          <div className={styles.header}>
+            <ModalAccessibleLabel>
+              <Text tag="h1" style="zen-heading-3" inline>
+                {title}
+              </Text>
+            </ModalAccessibleLabel>
+          </div>
+        </ModalHeader>
+        <ModalBody unpadded>
+          <div className={styles.body} dir={localeDirection}>
+            {children}
+          </div>
+        </ModalBody>
+        <ModalFooter
+          actions={footerActions}
+          appearance={type === "negative" ? "destructive" : "primary"}
+          automationId={automationId}
+        />
+      </div>
+    </GenericModal>
+  )
+}
 
 export default InputEditModal