+++ draft-packages/modal/KaizenDraft/Modal/Presets/RoadblockModal.tsx
@@ -1,4 +1,3 @@
-import classnames from "classnames"
 import * as React from "react"
 
 import { Heading } from "@kaizen/component-library"
@@ -15,7 +14,7 @@ import {
 
 import styles from "./RoadblockModal.scss"
 
-interface Props {
+export interface RoadblockModalProps {
   readonly isOpen: boolean
   readonly title: string
   readonly onDismiss: () => void
@@ -24,7 +23,7 @@ interface Props {
   readonly children: React.ReactNode
 }
 
-type RoadblockModal = React.FunctionComponent<Props>
+type RoadblockModal = React.FunctionComponent<RoadblockModalProps>
 
 const RoadblockModal = ({
   isOpen,
@@ -33,7 +32,7 @@ const RoadblockModal = ({
   dismissLabel = "Back",
   automationId,
   children,
-}: Props) => (
+}: RoadblockModalProps) => (
   <GenericModal
     isOpen={isOpen}
     onEscapeKeyup={onDismiss}
@@ -62,7 +61,7 @@ const RoadblockModal = ({
         </div>
       </ModalBody>
       <ModalFooter
-        actions={[{ label: dismissLabel, action: onDismiss }]}
+        actions={[{ label: dismissLabel, onClick: onDismiss }]}
         appearance="primary"
         automationId={automationId}
       />