+++ draft-packages/filter-menu-button/CHANGELOG.md
@@ -3,6 +3,71 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.10](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-filter-menu-button@1.1.9...@kaizen/draft-filter-menu-button@1.1.10) (2021-01-18)
+
+**Note:** Version bump only for package @kaizen/draft-filter-menu-button
+
+
+
+
+
+## [1.1.9](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-filter-menu-button@1.1.8...@kaizen/draft-filter-menu-button@1.1.9) (2021-01-18)
+
+
+### Bug Fixes
+
+* remove content wrapper for FilterMenuButton children ([#934](https://github.com/cultureamp/kaizen-design-system/issues/934)) ([92d223b](https://github.com/cultureamp/kaizen-design-system/commit/92d223b172ba1679de64afeec0e63326a2221e01))
+
+
+
+
+
+## [1.1.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-filter-menu-button@1.1.7...@kaizen/draft-filter-menu-button@1.1.8) (2021-01-14)
+
+**Note:** Version bump only for package @kaizen/draft-filter-menu-button
+
+
+
+
+
+## [1.1.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-filter-menu-button@1.1.6...@kaizen/draft-filter-menu-button@1.1.7) (2021-01-13)
+
+
+### Bug Fixes
+
+* update styling for FilterMenuButton metadata label ([#921](https://github.com/cultureamp/kaizen-design-system/issues/921)) ([16f2546](https://github.com/cultureamp/kaizen-design-system/commit/16f25465148561aeb98719eccb9a3f090e354a8d))
+
+
+
+
+
+## [1.1.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-filter-menu-button@1.1.5...@kaizen/draft-filter-menu-button@1.1.6) (2021-01-12)
+
+
+### Bug Fixes
+
+* remove the max-width to make the button expand with the metadata ([#920](https://github.com/cultureamp/kaizen-design-system/issues/920)) ([bb329ce](https://github.com/cultureamp/kaizen-design-system/commit/bb329ce5d87d3a3746141c77d72fc9cb91921c75))
+
+
+
+
+
+## [1.1.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-filter-menu-button@1.1.4...@kaizen/draft-filter-menu-button@1.1.5) (2021-01-11)
+
+**Note:** Version bump only for package @kaizen/draft-filter-menu-button
+
+
+
+
+
+## [1.1.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-filter-menu-button@1.1.3...@kaizen/draft-filter-menu-button@1.1.4) (2021-01-08)
+
+**Note:** Version bump only for package @kaizen/draft-filter-menu-button
+
+
+
+
+
 ## [1.1.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-filter-menu-button@1.1.2...@kaizen/draft-filter-menu-button@1.1.3) (2020-12-23)